Move Tests to its own asmdef to avoid possible compilation errors
because tests may use UnityEditor related APIs
