+++ .gitignore
@@ -1,0 +1,2 @@
+*.csproj
+*.csproj.meta
