+++ Runtime/Mathematics.Mathx.asmdef
@@ -3,7 +3,6 @@
     "rootNamespace": "",
     "references": [
         "GUID:d8b63aba1907145bea998dd612889d6b",
-        "GUID:343deaaf83e0cee4ca978e7df0b80d21",
         "GUID:2bafac87e7f4b9b418d9448d219b01ab",
         "GUID:2665a8d13d1b3f18800f46e256720795"
     ],
