+++ Tests/Mathematics.Mathx.Tests.asmdef
@@ -1,0 +1,22 @@
+{
+    "name": "Mathematics.Mathx.Tests",
+    "rootNamespace": "",
+    "references": [
+        "GUID:adfab0925da842c4db9a0d438ad877f7",
+        "GUID:d8b63aba1907145bea998dd612889d6b",
+        "GUID:343deaaf83e0cee4ca978e7df0b80d21",
+        "GUID:2bafac87e7f4b9b418d9448d219b01ab",
+        "GUID:2665a8d13d1b3f18800f46e256720795"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": true,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [
+        "UNITY_EDITOR"
+    ],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
